NBA Chord Diagram Generator/requirements.txt
pandas==1.1.2
nba-api==1.1.9
ttkthemes==3.1.1
pillow==7.2.0
chord==0.7.0